functionOC/requirements.txt
azure-functions==1.8.0
certifi==2021.10.8
numpy==1.22.2
numpyencoder==0.3.0
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
